refactor(mcp): split mcp_server.rs into module (mod.rs + params.rs)
- Split 1045-line mcp_server.rs into mcp_server/mod.rs (~950 lines) and mcp_server/params.rs (~150 lines)
- Satisfies coding standard: single file <= 1000 lines
- All 124 tests pass
